zvuvon-backend/requirements.txt
django
djangorestframework
django-cors-headers
requests
numpy
scipy